refactored_code2.0
