+++ index.html
@@ -77,7 +77,6 @@
       }
 
       window.onload = function () {
-    
         var baseurl =
           "https://raw.githubusercontent.com/RohiniDeshmukh/powerboost/main/src/";
           // "http://localhost:8000/src/";
